requirements.txt
annotated-types==0.6.0
anyio==4.2.0
asttokens==3.0.0
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
decorator==5.1.1
distro==1.9.0
exceptiongroup==1.2.0
executing==2.1.0
google-auth==2.25.2
google-auth-oauthlib==1.2.0
gspread==5.12.3
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
iniconfig==2.0.0
ipython==8.31.0
jedi==0.19.2
matplotlib-inline==0.1.7
oauthlib==3.2.2
openai==1.6.1
packaging==24.2
parso==0.8.4
pexpect==4.9.0
pillow==11.0.0
pluggy==1.5.0
prompt_toolkit==3.0.48
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==2.5.3
pydantic_core==2.14.6
Pygments==2.18.0
pyTelegramBotAPI==4.25.0
pytest==8.3.4
python-dotenv==1.0.0
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
sniffio==1.3.0
stack-data==0.6.3
telebot==0.0.5
tomli==2.2.1
tqdm==4.66.1
traitlets==5.14.3
typing_extensions==4.9.0
urllib3==2.1.0
wcwidth==0.2.13
